python/requirements.txt
websocket-client
python-dispatch
python-dotenv
